There is no need to hard code the settingPath
The setting path can be passed in the configuration (yaml) file. I remove
various lines that ended in spaces.
* timestream/manipulate/pipecomponents.py(ImageUndistorter, ColorCardDetector)
(ImageColorCorrector, TrayDetector, PotDetector, PlantExtractor): Remove
spaces ending lines.
* timestream/manipulate/pipeline.py(ImagePipeline): Removed settingPath from
the __init__ method. It should be in the configuration file.
(test): Removed test method. Was not needed and did not express the current
state of things.
the config file *.yml.
* util/pipeline_demo.py(settingFile, inputRootPath, outputRootPath): Changed
these variables to my local setup. These should not be set in code, but should
be gotten from a config file. Also removed settingPath from the ImagePipeline
instatiation line.