Initial commit for the segmentation logic
* util/pipeline_demo.py: Put the image output writing logic into a pipeline
component of its own.
* timestream/manipulate/pipecomponents.py (PlantExtractor): We will expect the
segmentation method name (defaults to k-means-square) and User defined
arguments. Args are defined in the yaml file.
(PlantExtractor.runReturns): We are returning the image and the ImagePotMatrix
instance.
(PlantExtractor.__init__): Use ChamberHandler to call the segmentation method
on the different pot centers.
(ResultingImageWriter): New pipeline component that will write any image to an
output file defined by an output timeseries instance defined in the context.
* timestream/manipulate/pipeline.py (complist): Added
* ResultingImageWriter.actName to the list of possible pipecomponents.
(process): Visualize should default to False.
* timestream/manipulate/plantSegmenter.py: New file containing logic for plant
and pot segmentation.
(ImagePot): New class that will represent individual pot images.
(ImagePot.rect): All pots will be represented as rectangles in an image. This
was a bit tricky as matrices in cv2 are indexed (height, width, :).
(ImagePot.maskedImage): Will return the segmented part of the image on top of
a black ([:,:,:]=0) background.
(ImagePotMatrix): New class to represent the matrix of pots in an image. For
now we assume that pots are in trays and trays fit in pictures. We might want
to come up with something more general in the future.
(ImagePotMatrix.asTuple): Method that will convert the ImagePotMatrix into a
list of lists so it can be serialiazable by json.
(ImagePotMatrix.getPot): Will allow individual pots to be indexed.
(ImagePotMatrix.ImageTray): New inner class that represents the trays.
(PotSegmenter): New parent class that will hold logic for the rest of the pot
segmenter approaches.
(PotSegmenter_KmeansSquare): Simple pot segmenter. Will use k-means to try to
separate two colors in the ImagePot.
(ChamberHandler): New class that brings everything (ImagePots, ImagePotMatrix,
PotSegmenter...) together under one roof.