Put segmentation logic in ImagePotHandler
* timestream/manipulate/plantSegmenter.py (ImagePotHandler.__init__): Add
PotSegmenter and ImagePotHandler as init arguments. The first is the one that
does the actual segmentation. The second is there for reference. Also add some
checks to the init method.
(ImagePotHandler.mask.setter): Remove method from class. Mask will only be
settable through the getter property.
(ChamberHandler.process): Add the segmenter object to each image pot handler.
This should happen in some other more intelligent way. We leave it like this
for now.