Add component to output to a csv file.
* timestream/manipulate/pipecomponents.py (ResultingFeatureWriter_csv): New
component that saves features to files with the names of the features. It will
put all the files in a directory specified by the user.
(ResultingFeatureWriter_ndarray.__init__): Raise error if path exists.
* timestream/manipulate/pipeline.py (complist): Add new component to the
component list
* timestream/manipulate/plantSegmenter.py (featureMethods): Change this to a
class method. So now the call is FeatureCalculator.featureMethods.
(ImagePotHandler.calcFeatures): Change the way we call featureMethods.