Use CIEL*a*b. Run k-means once.
CIEL*a*b is a bit better than RGB at noticing the differences in the green
spectrum. We have a pretty good idea of where the pot is, so we don't need to
grow it. And we add a texture feature explained in Massimo et. al.
(Image-based plant phenotyping with incremental learning and active contours).
* timestream/manipulate/plantSegmenter.py (PotSegmenter_KmeansSquare.segment):
Remove the logic that grows the pot in search of a better segmentation.
(PotSegmenter_KmeansSquare.getFeatures): New method that returns color and
texture features.