Add complexity calculation
We use complexity to decide if a mask is appropriate or not. If there is too
much complexity, it usually means that the segmentation when wrong.
* timestream/manipulate/plantSegmenter.py (calcComplexity): New method in
PotSegmenter_KmeansSquare that calculates complexity. We use Parrott et. al.
2008.
(segment): We check the relation of background -vs- foreground. If there is
more FG than BG we invert the mask.