Show segmentation when segmenting
We were not actually showing the segmentation before.
* timestream/manipulate/plantSegmenter.py (ImagePotHandler): We now always
return a mask. Its completely black if we have none available.
(ps): New property. This fixes a previously hidden bug.
(mask): When the mask contains negative 1 it has not been calcuated. Mask will
only be calculated if self._ps is not none and contains -1. If self._mask does
not contain -1, we return it
(rect): Reset self._mask to all -1 when changing self._rect.
(maskedImage): This property now returns the segmentated pot within the original
image. It's like a window into the segmentation.
(show): use the new maskedImage property to build a proper segmented image to
show.