Make previous pots available through previous mat.
* timestream/manipulate/pipecomponents.py (PlantExtractor.__call__): Get the
previous image pot matrix from context["ipm"]. At the end make the ipm
available by replacing it in context with self.ipm.
* timestream/manipulate/plantSegmenter.py (ImagePotMatrix.__init__): Add an
argument ipmPrev that will allow for pot continuity. Use the ipmPrev to add
iphPrev to each ImagePotHandler instantiation.
* util/pipeline_demo.py (context): Stop reinitilizing the context on every
image. This allows for pipecomponents to modify the context and re-use it on
popsterior runs.