Allow ResultingImageWriter to be inserted anywhere.
While ResultingImageWriter can be inserted anywhere, it will not be able to
create different output timestreams yet. We still need to add a way to define
multiple timestreams and a way to use them from inside the components.
* timestream/manipulate/pipecomponents.py (runExpects, runReturns): The
components that define their runReturns as [None] will be handled differently.
This allows for components to "tap into" the pipeline and gather information.
(__call__): These new type of components must return the same args that they
receive.
* timestream/manipulate/pipeline.py (__init__): We relax the checks on the new
component types. All the elements in the expecting list should be in the
providing list.