Add multiple output streams.
It is now possible to define and use different output streams from the yaml
configuration file. Each output streams has a name wich can be used in the
yaml config file.
* timestream/manipulate/pipecomponents.py (ResultingImageWriter.argNames): We
now expect the user to tell us to which output stream we should write the
images.
(__call__): The output steam is provided by the context.
* util/pipeline_demo.py (yfile): Contains the information from the yaml file.
(settings): Is yfile["pipeline"]
(outstreams): Is yfile["outstreams"]
(context): We now iterate through outstreams and create an output stream for
each one defined in the yaml config.