Expect an numpy.ndarray at startup.
This moves away from what we want in the end. To pass around TimeStreamImages
but it is necessary to be able to seamlessly start at any point in the
pipeline. When we do move towards TimeStreamImage, we should move all the
pipecomponents at once.
* timestream/manipulate/pipecomponents.py (runExpects): Expect a np.ndarray
instead of a TimeStreamImage.
(__call__): Don't access the pixels attribute.
* util/pipeline_demo.py (result): Initialize the pipeline with a np.ndarray.