Add getSegmented method for ImagePotHandler.
We do the mask property calculation in two steps, separating the actual
calculation from the assignment to self._mask. This is important because
the getSegmented code might run in parallel in different processes that
do not share memory
* timestream/manipulate/plantSegmenter.py (ImagePotHandler.getSegmented):
New method, replacing the "mask" property.
(ImagePotHandler.mask): Do basic checks and call getSegmented when we
(ImagePotHandler.mask.setter): Be very picky when we assign mask. Only
the correct type int64, the correct shape and ndarray can be set.
need the actual mask.
(PotSegmenter_KmeansSquare, ImagePotHandler) Make sure that the mask is
float64