Fix bug introduced in 1717e01
1717e01 put pot positions and tray positions in the output timestream metadata.
It would fail when `ts_out.image_data[parse.ts_format_date(date)]` was called
and the image was not there because it had not been written yet.

Relevant Traceback:
ceback (most recent call last):
  File "pipeline_demo.py", line 106, in <module>
    result = pl.process(context, [img.pixels], visualise)
  File "build/bdist.linux-x86_64/egg/timestream/manipulate/pipeline.py", line 82, in process
  File "build/bdist.linux-x86_64/egg/timestream/manipulate/pipecomponents.py", line 328, in __call__
KeyError: '2014_06_19_09_00_00'

The bug would only be seen on pipelines that called traydetector or potdetector
without calling resultingimagewriter.

* timestream/manipulate/pipecomponents.py (TrayDetector.__call__): Add relevant
value to context["outputwithimage"] so it can be output when the image is
written.
* timestream/manipulate/pipecomponents.py (PotDetector.__call__): Likewise
* timestream/manipulate/pipecomponents.py (ResultingImageWriter.__call__): Take
all values in context["outputwithimage"] and write them as metadata.
* util/pipeline_demo.py: Initializle context["outputwithimage"] to an empty dict.
* util/pipeline_demo_correct_images_daytime.py: Likewise