Correctly load image_data & new test case for this
We now correctly load ts.data and ts.image_data, and test that these are
loaded correctly. We also make ts.iter_by_whatever set the resulting
image.data to the data for that timepoint from ts.image_data, if that
timepoint is in the image_data dict. There's a new test case for this
too.