Remove "trays" from ImagePotMatrix
ImagePotMatrix will organize pots by ID, not by tray. The tray information can
still by added as an additional binding.
* timestream/manipulate/pipecomponents.py (PotDetector.__call__): Create the
ImagePotMatrix as soon as we detect the pots in PotDetector.
(PlantExtractor): We now expect a ImagePotMatrix object in the arguments.
* timestream/manipulate/plantSegmenter.py (ImagePotHandleriphPrev): We put this
in a property and use _iphPrev internally.
(ImagePotHandler.setSbind): Allows the creation of an arbitrary binding for each
pot.
(ImagePotMatrix): We remove the tray class and have a simple dictionary of pots
indexed by the potId.
(ImagePotMatrix.addPot): New method that adds pots to the matrix dictionary.