Work on internal consistency of TimeStreamImage
TimeStreamImage is driven by _path. _path should be valid at end of all
methods.!! In general we only use property variables inside methods when we have
a very good reason to do so.
* timestream/__init__.py (TimeStreamImage.write): Use _pixels and _path instead
of pixels and path.
(parent_timestream):Special case where we want to leave TimeStreamImage without
a _timestream. When ts==None, we leave evertying (_*) untouched
(datetime): Reset _path and _pixels when we set datetime.