_pixels orthogonal from _path, _datetime and _timestream
TimeStreamImage was trying to keep _pixels, _path, _datetime and _timestream
consistent. However, _pixels is orthogonal from the other three; we need to be
able to set _pixels without changing _path and viseversa. This commits removes
the checks to _pixels.
* timestream/__init__.py (TimeStreamImage.read): Change method to force a
refresh in _pixels. Using this method is a way to replace a _pixels by
forcing a reading of _path.
(TimeStreamImage.pixels): Removed some logic and put it in the read method. We
call the read method from here.
(_pixel): Remove checks of consistency between _pixels and _path