Add time stream list.
* util/derandui.py: Add GPL text.
(timeStreams): Dictionary that holds all the opened timestreams.
(onClickTimeStreamList): When we click on the "+" symbol we add a timestream
directory. We fail with an error message is TimeStream.load fails.
* util/derandui.ui: We add the timestream list as a TableWidget to the user
interface. We add a hidden colomn to relate the rows in the timestream list to
the new self.timeStreams dictionary.