Open TimeStreamTraverser and show an image
* timestream/__init__.py (TimeStreamTraverser): New class that inherits from
TimeStream. It implements the timestream as a backwards and forwards moving
object. The rest is basically the same as TimeStream.
(TimeStreamTraverser.{next,prev,curr}): New methods for new class. They are
responsable for keeping internal track of where the timestreamtraverser is. If
we overflow we continue at end or start of time stream.
* util/derandui.py (showImage): Puts an image on the scene.
(selectFile): Removed this method.