Remove specifics from configuration.py
* timestream/manipulate/configuration.py (PCFGConfig.__init__): Use the
underscore "_" for internal variables so we can have subsections named
"configFile" and "depth".
(PCFGConfig.loadFromYaml): Configuration is just a structure that holds whatever
is on the path that it is handed. It reflects what is in a config file or what
is added through the methods. We should not add stuff that the caller does not
know about.
* util/pipeline_demo.py (plConf): We can add the confFile subsection after we
create plConf and it should have the same result.