Populate first column when adding a timestream
* timestream/manipulate/pot.py(numPots): New property that returns the number of
pots in a ImagePotMatrix.
* util/derandui.py (WindowTS): New class that contains the relation between
TimeStreamTraverser instance and the csv table. Its easier to have them together
as they change at the same time in the GUI.
(onClickTimeStreamList): Check to see if TimeStreamTraverser has all relevant
information before we add to list. We use the new WindowTS here. Ignore clicks
to the last row in the tslist.
(selectRowTS): Incorporate ddm in when selecting row in tslist.
(DropDownMenu, DropDownMenu_TS): New classes that contain menu widget for csv
table headers. They will contain all logic that has to do with filling the
column under the header.
* util/derandui.ui: Initialize the csv with 2 columns and one row, No headers.