Reorder Col1 based on Col0
* util/derandui.py (FirstTwoColums.E): String "--empty--" to put in empty rows
to identify them
(FirstTwoColums._num0Rows): Number of rows for column 0. This is different from
the number of rows in the table.
(refreshCol{0,1}Header): Use method setCurrentIndex to initiate refreshCol{0,1}
calls.
(refreshCol1): Skip the empty rows when copying to column 1
(colsRebind): This will refresh the relation between col0 and col1. Col0 will
not be reorder, but col1 will be reordered (together with the rest of the
columns 2...end) to have a 1-to-1 relation with col0. If we find no relation we
insert an empty row.
(swapRow): By default we fill the cell with (" ")