Refactor BindingTable's column refresh logic
The columns now need to be refreshed in a specific order. 0->1->2. This is done
to be able to reorder all the csv information in a efficient way.
* util/derandui.py (BindingTalbe.imgRefresh): Don't print anything for now.
(BindingTable._colActionDispatcher): New method that anlyzes what type of change
is happening and calls the refresh methods in the required order.
(BindingTable.refreshCol{0,1,2}Header): To these methods we added a blockSignal
call in order to avoid signals firring while we do changes in the comboboxes.
(Bindingtable._refreshCol{0,1,2}): We move some of the logic in the removed
colsRebind method into these methods. It is important to call them in the
correct order 0>1>2.
(BindingTable._blankCol): New helper method to blank columns.
(BindignTable._copyCol): New helper method to copy one column to another offset.