Keep track of the TS combo box offsets.
This allows us to remember what the user chose and get all the corresponding
data once we start the derandomization.
* util/derandui.py (DerandmizeGUI._activeTS): is now the selected offset in the
TS list.
(Second Row in TSlist): Now contains a 3-tuple as data. (TS, items, offset). TS
is timestream object, items is the list of possible mids related to TS and
offset is the selected mid.