Use metaids from TimeStream settings.
* timestream/__init__.py (next, prev): Actually add and subtract from offset
when we next and prev.
(pickleload): When we load we just want to read.
* timestream/manipulate/pipecomponents.py (PopulatePotMetaIds): Remove class.
We get the metaids from the TS now.
(runExpects, runReturns):
* timestream/manipulate/pipeline.py (complist): Remove the PopulatePotMetaIds
class from possible pipecomponents.
* util/derandui.py: Get our metid information from TimeStream settings.
* util/pipeline_demo.py: Change the way we merge the configuration files. We add
all pipeline related information to a pipeline component if we find a pipelin
name. If we don't fine a name, we add it to config.general.