Add derandomization logic.
We got the code to a point where it actually creats a de-randomized image. It is
slow due to lots of unnecessary loops. It is also very memroy hungry.
We now use a deransStructure of the form {mid:{TSpath:[pot,pot...],
TSpath;[pot,pot...]}, mid:{TSpat:[pot,pot,...]}}. This allows us to express what
is needed for the derandomization.

* timestream/manipulate/pipecomponents.py(DerandomizeTimeStreams): New class
component that houses all the logic that creats a derandomized timestream image.
(DerandomizeTimeStreams.getMidGrpImg): New method that creats the image for a
meta id group of pots.
(DerandomizeTimeStreams.createCompoundImage): New method that returns the
derandomized image.
(DerandomizeTimeStream.createMids): New method that creats an intermediate
structure of the form {mid:[pot, pot,....], mid:[pot,pot,pot]}. We are using new
TimeStreams for this which might be the reason for the increased memory usage.
* timestream/manipulate/pot.py: Various small non-related fixes.
* util/derandui.py(DerandomizeGUI._derandomize): New method that sets variables
for the execution of the derandomization pipeline. It has the task of creating
the derandStruct.
* timestream/__init__.py (TimeStreamTraverser.timestamps): New property method
that returns all valid timestreams.
(TimeStreamTraverser.getImgByTimeStamp): New method that returns an image based
on a timestamp.
* util/derandui.ui: Add a new button for initializing the derandomization.