Output time stream name is now settable.
* util/pipeline_demo.py: By default the output time stream suffix is the same as
the input time stream name. This created unwanted behavior when the last part of
the input time stream appeared in the output name. This can now be circumvented
by adding specifying the root name of the timestream by using "tsdirrootname" in
the timestream.yml config file. "tsdirrootname: BVZ0036-GC02L-C01~fullres" would
be a valid line.