Rename pipeline_demo.py and move to a scripts dir
It's not really a demo anymore, it's a fully fledged entry point into
the pipelines. This name refects this maturity.

Also, util/ was supposed to be for code that is just for internal use by
developers, so it would be good if we moved all the rest of the stuff
out to ./scripts or wherever we end up. Maybe ./bin.