Propagate exceptions throughout the pipecomponents
When a component throws an exception we complete the pipeline by passing the
exception object instead of the normal return val. By default each component
will re-raise the exception. Components can intervene in the default behavior by
overriding the __chkExcept__ method.
* scripts/run-pipeline.py: We use the TimeStreamTraverser instead of TimeStream.
This allows for more control of the iteration. We now have access to the
timestamp that generates the next img instance.
* timestream/__init__.py (TimeStreamTraverser.__init__): New input argument
live_except. If true we check validity of img every time we try to
generate/access it.
(TimeStreamTraverser.curr): Now calls getImgByTimeStamp to execute command
(TimeStreamTraverser.getImgByTimeStamp): Checks for errors when live_except is
True. Note that when live_except is false we check for errors in __init__.
* timestream/manipulate/__init__.py: Moved all exception classes here.
(PCExBadConfig): New exception to be thrown when a configuration element is not
right.
(PCExBadContext): Thrown when a context element is not right.
* timestream/manipulate/pipecomponents.py (PipeComponent): Before we call
__exec__, we check to see if there is an exception and re-raise it.
(PipeComponent.__chkExcept__): New method that is called to check for existance
of exception in input args. This is the method that should be overloaded in case
a component needs to do something with the exception.
* timestream/manipulate/pipeline.py (process): We re-raise the exception that
was returned after running the whole pipeline.